+++ examples/Zephyr/Example1_GetPositionAndTime_Zephyr/prj.conf
